pyproject.toml
[project]
description = "A starter template for marimo notebooks with uv"
name = "marimo-template"
requires-python = ">=3.12"
version = "0.1.0"

[dependency-groups]
dev = [
  "pre-commit>=4.0.1",
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.3",
  "marimo>=0.13.14",
  # vault library
  "pyyaml>=6.0.2",
  "networkx>=3.4",
  "numpy>=2.0.0",
  "altair>=5.5.0",
  "polars>=1.0.0",
  "duckdb>=1.2.0",
  "pyarrow>=19.0.0",
  "httpx>=0.28.0",
  # UI testing
  "playwright>=1.50.0",
  "pytest-playwright>=0.7.0",
  "requests>=2.32.0",
]

[tool.pytest.ini_options]
pythonpath = ["./src"]
testpaths = ["tests"]
# Use -m e2e to run browser tests; they are excluded by default
addopts = [
  "--cov=src",
  "--cov-report=term-missing",
  "-m",
  "not e2e",
]
markers = [
  "e2e: End-to-end tests that require a live Marimo server and Playwright browsers",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
]

[tool.marimo.ai]
rules = "- prefer polars over pandas\n- make charts using altair"

[tool.marimo.save]
autosave = "after_delay"
autosave_delay = 1000
format_on_save = true

[tool.marimo.package_management]
manager = "uv"

[tool.marimo.runtime]
pythonpath = ["./src"]
